requirements.txt
Django
django-mathfilters
requests
Levenshtein
python-whois
dnspython
pyquery
numpy
waybackpy
shodan
scikit-learn
selenium
pandas
tldextract
bs4
Pillow
# plotly
gdown
# kaleido
webdriver-manager
django-redis
django-debug-toolbar